+++ src/audiofilters/Filter.h
@@ -12,6 +12,7 @@
 #include "py/obj.h"
 
 #include "audiocore/__init__.h"
+#include "audiofilters/__init__.h"
 #include "synthio/Biquad.h"
 #include "synthio/__init__.h"
 #include "synthio/block.h"
@@ -20,12 +21,8 @@ extern const mp_obj_type_t audiofilters_filter_type;
 
 typedef struct {
     audiosample_base_t base;
-    mp_obj_t filter;
+    audiofilters_filter_chain_t filter;
     synthio_block_slot_t mix;
-
-    mp_obj_t *filter_objs;
-    size_t filter_states_len;
-    biquad_filter_state *filter_states;
 
     int8_t *buffer[2];
     uint8_t last_buf_idx;